pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "annlib"
version = "0.1.0"
description = "대학원 수업용 논문 관리 올인원 대시보드"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["research", "papers", "obsidian", "rag", "notebooklm"]

dependencies = [
    "streamlit>=1.32",
    "playwright>=1.44",
    "pypdf2>=3.0",
    "requests>=2.31",
    "click>=8.1",
    "rich>=13.0",
    "openai>=1.0",
    "google-genai>=1.0",
    "anthropic>=0.25",
    "python-dotenv>=1.0",
    "aiofiles>=23.0",
    "httpx>=0.27",
]

[project.scripts]
annlib = "annlib.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["annlib"]
